Fixes #34554 - remove the improper validation
During the transition to the openscap_parser gem to avoid using the
ffi wrappers around the C library, a trivial validation was introduced.
Such validation turned to be very limited and not supporting non-RH
valid SCAP content. This patch removes the validation. The only proper
solution would be to shell out to the oscap tool, but the purpose was to
be independent of it.