Fixes #37554 - Move most of request processing into its own fork
Ruby is notoriously bad with giving back memory it allocates. With
reports being a couple of MBs bzip2'ed and tens of megabytes
decompressed, memory consumption has a tendency to skyrocket as reports
get processed.

This change forks the process before report is processed and the
entirety of the report processing happens in the fork. Once the
processing is done, the forked process exits and all the memory it might
have allocated is released.