Fixes #38224 - Make reports available after upload timeout
Currently, when the OpenSCAP scan report is not uploaded to Foreman
after the scan finishes because it times out, the report is saved in
/var/spool on the proxy. When the report is requested from foreman, it
cannot be found because it is not being searched for in spool.

This PR aims to fix this by searching for the report in the spooldir and
moving it to the proper storage place.